opv_import/makelot/utils.py
# coding: utf-8

# Copyright (C) 2017 Open Path View, Maison Du Libre
# This program is free software; you can redistribute it and/or modify
# it under the terms of the GNU General Public License as published by
# the Free Software Foundation; either version 3 of the License, or
# (at your option) any later version.
# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
# GNU General Public License for more details.
# You should have received a copy of the GNU General Public License along
# with this program. If not, see <http://www.gnu.org/licenses/>.

# Contributors: Benjamin BERNARD <benjamin.bernard@openpathview.fr>
# Email: team@openpathview.fr
# Description: Utils for makelot.

def bit_len(int_type: int) -> int:
    """
    Returns int_type length, position of last non 0 bit.

    :param int_type: An int value.
    :type int_type: int
    :return: position of last non 0 bit.
    :rtype: int
    """
    length = 0
    while (int_type):
        int_type >>= 1
        length += 1
    return(length)
